more restructure stuff
