+++ tests/tests-integrations/src/main/java/org/eclipse/dirigible/integration/tests/api/rest/EnabledMultitenantModeIT.java
@@ -10,7 +10,7 @@
 package org.eclipse.dirigible.integration.tests.api.rest;
 
 import org.eclipse.dirigible.commons.config.DirigibleConfig;
-import org.eclipse.dirigible.tests.framework.ide.DirigibleTestTenant;
+import org.eclipse.dirigible.tests.framework.tenant.DirigibleTestTenant;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
