+++ tests/tests-integrations/src/main/java/org/eclipse/dirigible/integration/tests/ui/tests/MultitenancyITTestProject.java
@@ -16,13 +16,13 @@ import org.eclipse.dirigible.tests.framework.awaitility.AwaitilityExecutor;
 import org.eclipse.dirigible.tests.framework.browser.Browser;
 import org.eclipse.dirigible.tests.framework.browser.BrowserFactory;
 import org.eclipse.dirigible.tests.framework.browser.HtmlElementType;
-import org.eclipse.dirigible.tests.framework.ide.DirigibleTestTenant;
 import org.eclipse.dirigible.tests.framework.ide.EdmView;
 import org.eclipse.dirigible.tests.framework.ide.IDE;
 import org.eclipse.dirigible.tests.framework.ide.IDEFactory;
 import org.eclipse.dirigible.tests.framework.logging.LogsAsserter;
 import org.eclipse.dirigible.tests.framework.projects.BaseMultitenantTestProject;
 import org.eclipse.dirigible.tests.framework.restassured.RestAssuredExecutor;
+import org.eclipse.dirigible.tests.framework.tenant.DirigibleTestTenant;
 import org.eclipse.dirigible.tests.framework.util.ProjectUtil;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -103,10 +103,8 @@ class MultitenancyITTestProject extends BaseMultitenantTestProject {
     }
 
     /**
-     * Verifies indirectly:<br>
-     * - MultitenancyIT/views/readers.view is created and it is working<br>
-     * - MultitenancyIT/csvim/data.csvim is imported <br>
-     * - default DB datasource is resolved correctly
+     * Verifies indirectly:<br> - MultitenancyIT/views/readers.view is created and it is working<br> - MultitenancyIT/csvim/data.csvim is
+     * imported <br> - default DB datasource is resolved correctly
      */
     private void verifyView(DirigibleTestTenant tenant) {
 
@@ -119,17 +117,12 @@ class MultitenancyITTestProject extends BaseMultitenantTestProject {
                              .body("[0].READER_FIRST_NAME", equalTo("Ivan"))
                              .body("[0].READER_LAST_NAME", equalTo("Ivanov"))
                              .body("[1].READER_FIRST_NAME", equalTo("Maria"))
-                             .body("[1].READER_LAST_NAME", equalTo("Petrova")),
-                25);
+                             .body("[1].READER_LAST_NAME", equalTo("Petrova")), 25);
     }
 
     /**
-     * Verifies indirectly:<br>
-     * - edm generated schema is created<br>
-     * - generated REST is created and it works<br>
-     * - topic listener works<br>
-     * - job has been executed<br>
-     * - default DB datasource is resolved correctly
+     * Verifies indirectly:<br> - edm generated schema is created<br> - generated REST is created and it works<br> - topic listener
+     * works<br> - job has been executed<br> - default DB datasource is resolved correctly
      */
     private void verifyEdmGeneratedResources(DirigibleTestTenant tenant) {
         restAssuredExecutor.execute(tenant, () -> verifyBookREST(tenant));
@@ -165,8 +158,9 @@ class MultitenancyITTestProject extends BaseMultitenantTestProject {
     }
 
     private void verifyJobExecuted(DirigibleTestTenant tenant) {
-        String expectedMessage = "Job: found [1] books. Books: [[{\"Id\":1,\"Title\":\"Title[" + tenant.getName()
-                + "]\",\"Author\":\"Author[" + tenant.getName() + "]\"}]]";
+        String expectedMessage =
+                "Job: found [1] books. Books: [[{\"Id\":1,\"Title\":\"Title[" + tenant.getName() + "]\",\"Author\":\"Author["
+                        + tenant.getName() + "]\"}]]";
         verifyMessageLogged(expectedMessage, testJobLogsAsserter);
     }
 
@@ -178,18 +172,15 @@ class MultitenancyITTestProject extends BaseMultitenantTestProject {
     }
 
     private void verifyListenerExecuted(DirigibleTestTenant tenant) {
-        String expectedMessage = "Listener: found [1] books. Books: [[{\"Id\":1,\"Title\":\"Title[" + tenant.getName()
-                + "]\",\"Author\":\"Author[" + tenant.getName() + "]\"}]]";
+        String expectedMessage =
+                "Listener: found [1] books. Books: [[{\"Id\":1,\"Title\":\"Title[" + tenant.getName() + "]\",\"Author\":\"Author["
+                        + tenant.getName() + "]\"}]]";
         verifyMessageLogged(expectedMessage, eventListenerLogsAsserter);
     }
 
     /**
-     * Verifies indirectly:<br>
-     * - MultitenancyIT/tables/reader.table is created<br>
-     * - MultitenancyIT/csvim/data.csvim is imported <br>
-     * - MultitenancyIT/odata/readers.odata is configured <br>
-     * - OData is working<br>
-     * - default DB datasource is resolved correctly
+     * Verifies indirectly:<br> - MultitenancyIT/tables/reader.table is created<br> - MultitenancyIT/csvim/data.csvim is imported <br> -
+     * MultitenancyIT/odata/readers.odata is configured <br> - OData is working<br> - default DB datasource is resolved correctly
      */
     private void verifyOData(DirigibleTestTenant tenant) {
         restAssuredExecutor.execute(tenant, () -> {
